+++ packages/datatrak-web/src/components/UnsyncedDataGuard.tsx
@@ -1,0 +1,48 @@
+import React, { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+
+import { useHasUnsyncedDataQuery } from '../api/queries';
+import { ROUTES } from '../constants';
+import { useNavigationBlocker } from '../utils';
+import { ConfirmationModal } from './ConfirmationModal';
+
+/**
+ * Blocks navigation to /logout when there is unsynced data and shows a
+ * confirmation modal. If the user confirms, navigation proceeds to the logout
+ * page. If cancelled, navigation is reset and the user is sent home.
+ */
+export const UnsyncedDataGuard = () => {
+  const { data: hasUnsyncedData } = useHasUnsyncedDataQuery();
+  const navigate = useNavigate();
+
+  const [isOpen, setIsOpen] = useState(false);
+
+  const { proceed, reset } = useNavigationBlocker({
+    active: !!hasUnsyncedData,
+    onBlock: () => setIsOpen(true),
+    shouldBlock: (pathname: string) => pathname === ROUTES.LOGOUT,
+  });
+
+  const handleClose = () => {
+    setIsOpen(false);
+    reset();
+    navigate(ROUTES.HOME);
+  };
+
+  const handleConfirm = () => {
+    setIsOpen(false);
+    proceed();
+  };
+
+  return (
+    <ConfirmationModal
+      isOpen={isOpen}
+      onClose={handleClose}
+      onConfirm={handleConfirm}
+      heading="Unsynced data"
+      description="You are about to log out with unsynced data! Go back to your home page and sync using the top right sync button and sync before logging out"
+      confirmLabel="Log out anyway"
+      cancelLabel="Stay logged in"
+    />
+  );
+};
