+++ packages/datatrak-web/src/utils/index.ts
@@ -10,6 +10,7 @@ export {
 } from './sizeClasses';
 export { errorToast, infoToast, successToast } from './toast';
 export { useBeforeUnload } from './useBeforeUnload';
+export { useNavigationBlocker } from './useNavigationBlocker';
 export { useHasVideoInput } from './useHasVideoInput';
 export { useFromLocation } from './useLocationState';
 export { formatEntityForResponse, formatEntitiesForResponse } from './formatEntity';
